Add files via upload
The liver is a critical organ in the human body that is responsible for an array of functions that help support metabolism, immunity, digestion, detoxification, vitamin storage among other functions. It comprises around 2% of an adult's body weight. This is a Machine learning project which makes use of classification algorithm to determine occurrences of liver diseases amongst different age bracket and gender.